requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.3
pydantic>=1.8.2
mcp-server>=0.1.0 